New Redirect
To the very small amount of people that actually keep up with this
It has been a long time since I have changed this repository. I have had a lot of school work (Im in the IB program), but im back. I am now testing for security, specifically the login page.